Separate ApplicationConfigurationReaderTests into own file
